bin/nvh: Add "forced on" state for xz
Skips node version checks, which would otherwise gate xz usage.

In other words:
Overrides any checks that would normally be run,
so that nvh always attempts to use xz.

---

Also updates "--use-xz" to not set the "xz forced on" state.
(This matches the flag's description in the help messages.)

---

Partly adapted from this comment (with liberties taken on implementation):
https://github.com/shadowspawn/nvh/issues/8#issuecomment-533034644

Co-authored-by: John Gee <john@ruru.gen.nz>